refactor(ui): remove obsolete shared folder, ServerErrorBanner moved to ui/feedback
